Add seek function to FLACDecoder
Initial version of seek function in FLACDecoder class takes a requested
(absolute) sample position in the file and attempts to move the input
stream to the frame which contains the requested sample position.
Returns the actual sample position from which the next frame will be
read (using readNextFrame).

Precision is at the frame level due to the design of the FLACDecoder
class. Sample level precision would require some internal buffering and
so is left to the client application to handle.

Required minor additions to BitInputStream and RandomFileInputStream
classes.

Also, minor typo corrected in StreamInfo toString().